+++ LookupAnything/GameHelper.cs
@@ -202,9 +202,12 @@ namespace Pathoschild.Stardew.LookupAnything
             if (!npc.isVillager())
                 return false;
 
-            return this.Metadata.Constants.ForceSocialVillagers.TryGetValue(npc.Name, out bool social)
-                ? social
-                : npc.CanSocialize;
+            if (this.Metadata.Constants.ForceSocialVillagers.TryGetValue(npc.Name, out bool social))
+                return social;
+
+            return
+                Game1.player.friendshipData.ContainsKey(npc.Name)
+                || npc.CanSocialize;
         }
 
         /// <summary>Get how much each NPC likes receiving an item as a gift.</summary>