+++ LookupAnything/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed some untranslated text.
+* Fixed NPC lookups not showing social info for some custom mod NPCs. (This affects NPCs with friendship data but `CanSocialize` set to false.)
 * Improved translations. Thanks to spindensity (updated Chinese)!
 
 ## 1.28.3